Move all work to main branch
This commit serves as the root commit for the main branch. History is, sadly,
being purposefully removed due to the large amount of files present in the
previous default branch. The previous branch still exists, so the appropriate
authors of commits can be credited, but the branch has been archived. To look
up commits prior to this change, see the archived "master" branch.